+++ pycm/pycm_profile.py
@@ -3,6 +3,8 @@
 from pycm import *
 import numpy as np
 
+np.random.seed(123)
+
 x1 = np.random.randint(low=0, high=400, size=2000000)
 x2 = np.random.randint(low=0, high=400, size=2000000)